# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Be Health",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:82/api",
            "description": "Laravel Swagger API Server"
        }
    ],
    "paths": {
...
        "/treatment-amount/{treatment_id}": {
            "get": {
                "tags": [
                    "Kit"
                ],
                "summary": "Get kit amount",
                "description": "Retrieve the amount of a kit associated with a specific treatment",
                "operationId": "213ddd92c4278b18f12f009a2021e65d",
                "parameters": [
                    {
                        "name": "treatment_id",
                        "in": "path",
                        "description": "ID of the treatment",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1.23
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Treatment or kit not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
...
}